[plock-17]: fix global shortcuts via load settings
